+++ sphinx/application/data/features/feature-repository/build.gradle
@@ -25,7 +25,7 @@ dependencies {
     api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-media')
     api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-message')
     api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-subscription')
-    api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-podcast')
+    api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-feed')
 
     api project(path: ':sphinx:application:network:concepts:concept-meme-input-stream')
     api project(path: ':sphinx:application:network:concepts:queries:concept-network-query-meme-server')
