+++ sphinx/screens-detail/join-tribe/join-tribe/build.gradle
@@ -42,6 +42,7 @@ dependencies {
     implementation project(path: ':sphinx:application:data:concepts:concept-media-cache')
     implementation project(path: ':sphinx:application:data:concepts:repositories:concept-repository-contact')
     implementation project(path: ':sphinx:application:data:concepts:repositories:concept-repository-chat')
+    implementation project(path: ':sphinx:application:data:concepts:repositories:concept-repository-feed')
     implementation project(path: ':sphinx:application:network:concepts:queries:concept-network-query-chat')
     implementation project(path: ':sphinx:application:common:menus:menu-bottom-picture')
     implementation project(path: ':sphinx:screens-detail:common:detail-resources')
