+++ sphinx/screens-detail/podcast-player/podcast-player/build.gradle
@@ -44,7 +44,7 @@ dependencies {
   implementation project(path: ':sphinx:application:data:concepts:repositories:concept-repository-chat')
   implementation project(path: ':sphinx:application:data:concepts:repositories:concept-repository-message')
   implementation project(path: ':sphinx:application:data:concepts:repositories:concept-repository-contact')
-  implementation project(path: ':sphinx:application:data:concepts:repositories:concept-repository-podcast')
+  implementation project(path: ':sphinx:application:data:concepts:repositories:concept-repository-feed')
   implementation project(path: ':sphinx:application:network:concepts:queries:concept-network-query-chat')
 
   api project(path: ':sphinx:service:concepts:concept-service-media-player')
