+++ sphinx/screens-detail/video-screen/video-screen/build.gradle
@@ -46,6 +46,7 @@ dependencies {
     implementation project(path: ':sphinx:application:data:concepts:concept-image-loader')
     implementation project(path: ':sphinx:application:data:concepts:concept-relay')
     implementation project(path: ':sphinx:application:data:concepts:repositories:concept-repository-chat')
+    implementation project(path: ':sphinx:application:data:concepts:repositories:concept-repository-feed')
 
     implementation deps.androidx.lifecycle.hilt
     implementation deps.google.hilt
