+++ sphinx/screens/chats/chat-common/chat-common/build.gradle
@@ -57,7 +57,7 @@ dependencies {
     api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-contact')
     api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-message')
     api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-media')
-    api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-podcast')
+    api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-feed')
 
     implementation project(path: ':sphinx:application:network:concepts:clients:concept-network-client-crypto')
