+++ sphinx/screens/dashboard/dashboard/build.gradle
@@ -40,7 +40,7 @@ dependencies {
     // Sphinx
     api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-contact')
     api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-chat')
-    api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-podcast')
+    api project(path: ':sphinx:application:data:concepts:repositories:concept-repository-feed')
     api project(path: ':sphinx:application:common:wrappers:wrapper-chat')
 
     implementation project(path: ':sphinx:activity:insetter-activity')
